Move MapShed tasks, calcs into a module
Since MapShed will require a number of Celery tasks to interface
with the geoprocessing service, it is prudent to collect them into
their own module.
