+++ src/mmw/apps/modeling/mapshed/data/sample_input.gms
